requirements.txt
django==4.0
django-admin-rangefilter==0.9.0
django-baton==2.3.0
psycopg2-binary==2.9.3
pyOpenSSL==22.0.0
